% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refineBasal.R
\name{refineBasal}
\alias{refineBasal}
\title{Refine basal-like muscle invasive bladder cancer}
\usage{
refineBasal(
  expr = NULL,
  isBasal = NULL,
  res.path = getwd(),
  res.name = "REFINED_MIBC_BASAL"
)
}
\arguments{
\item{expr}{A numeric expression matrix with row features and sample columns.}

\item{isBasal}{A logic vector to indicate if the sample is basal-like or not. NULL by default and all samples will be assumed as basal-like.}

\item{res.path}{A string value to indicate the path for saving the prediction result.}

\item{res.name}{A string value to indicate the name of the output prediction table.}
}
\value{
A data.frame storing the random forest prediction results, including `samID` for sample name, `prob` for probability as basal-inflamed, and `basal` for final classifications using 0.5 as cutoff.
}
\description{
Refine pre-identified basal-like muscle invasive bladder cancer into either basal-inflamed or basal-noninflamed subtype based on a random forest classifier.
}
\examples{
library(refineMIBC)
load(system.file("extdata", "demo.RData", package = "refineMIBC", mustWork = TRUE)) # load example data
expr <- demo$MIBC.expr
subt <- demo$MIBC.subt
isbasal <- subt$TCGA == "Basal_squamous"
iBasal  <- refineBasal(expr    = expr,
                       isBasal = isbasal)
head(iBasal)
}
