% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{templates}
\alias{templates}
\title{Template for classifying MIBC using NTP}
\format{

}
\usage{
templates
}
\description{
A data frame storing the subtype-specific markers for MIBC.
}
\keyword{datasets}
