% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{rfClassifier}
\alias{rfClassifier}
\title{Random forest classifier for refining basal-like MIBC}
\format{

}
\usage{
rfClassifier
}
\description{
An R object derived from varSelRF.
}
\keyword{datasets}
