% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{templates}
\alias{templates}
\title{Template for classifying MIBC using NTP}
\format{
A data frame with 120 rows (30 markers for each subtype) and 3 variables.
\describe{contains templates the subtype-specific up-regulated markers for MIBC.}
}
\usage{
templates
}
\description{
A data frame storing the subtype-specific markers for MIBC.
}
\keyword{datasets}
